@pitdicker 2017-12-17
> Not sure what you mean about not liking the way it works though?

Some parts seem not very ergonomic. For example the `new` function takes an already existing RNG. Why does it need you to initialize the wrapped rng separately? If it knows how to reseed an rng, surely it can also initialize one?

The `from_reseeder` function does not make much sense to me. The main difference with `new` is that it comes with a default threshold. The `DEFAULT_RESEEDING_THRESHOLD` is very small and basically always wrong. The threshold value depends on the paranoia of the application/library, and on the wrapped rng, so I don't thing having a default is a good idea. Also it takes a fixed seed, instead of using the `SeedFromRng` trait.

And I didn't like the `Reseeder` trait. Why can't `ReseedingRng` just use some other `Rng`, without requiring a custom wrapper. That were my thoughts yesterday though, I may be coming around on the last issue...