@pitdicker 2017-12-17
I have been trying out a wild idea today to reduce the overhead of `ThreadRng`, which uses `ReseedingRng`. Because that seems like the commonly used interface, I don't really want it to look bad :-). In `rand` in the nursery its performance is only 50% of the RNG it wraps. After sprinkling some `[inline]`'s and removing some wrappers it gets to 80% of the RNG.

Is seems sensible to assume that if you want to reseed an RNG, it will probably be a cryptographic RNG, not some simple one. The variants I have seen so far all generate blocks of results, not one result at a time. One way to reduce the overhead of `ReseedingRng` is to do the reseed 'bookkeeping' only when a new block of results is generated, not with every `next_*`.

As a test I have added an `RngCore` trait:
```rust
pub trait RngCore: Sized {
    type Results: AsRef<[u32]>;

    fn init(seed: &[u32]) -> Self;
    fn generate(&mut self, results: &mut Self::Results);
    fn results_empty() -> Self::Results;
}
```

It just exposes the core algorithm form an RNG, and does not include the buffering etc necessary to implement the `Rng` trait. It does not really make the implementation of HC-128 less clean, and should also not be to hard too implement for ISAAC and ChaCha.

Because `ReseedingRng` now has access to the RNG's algorithm, it can only use that part and implement it's own buffering and bookkeeping. This should (not completely true for `next_u32` yet) bring the overhead of `ReseedingRng` down to almost 0%, and bring the performance of `ThreadRng` within 90% of the wrapped RNG.

Do you think this a direction worth pursuing?

This is my current super ugly, many things comment out, WIP https://github.com/pitdicker/rand/commit/25dfbdddbc13cb3c59c865cdecd529f3d0784781