@dhardy 2017-12-18
This trait requires the implementation to use `u32` internally though, right? Can that be made a parameter? Also, I don't much like using a `&[u32]` seed; maybe you can do something like:

```rust
pub trait<T> RngCore: SeedableRng {
    type Results: AsRef<[T]>;

    fn init(seed: &<Self as SeedableRng>::Seed) -> Self;
```

Not quite sure what I think right now; this would make `ReseedingRng` only usable for certain classes of RNGs, right? But maybe that's an advantage allowing `reseed` to combine both current state (or output buffer) with a fresh seed instead of simply replace with a new seed. You're right, the wrapper is not very useful for fast, weak PRNGs.