@pitdicker 2017-12-18
I was already a little proud I got the extra abstraction working, including the Asref slice trick to be generic over arrays. But what you write is much better!

> But maybe that's an advantage allowing `reseed` to combine both current state (or output buffer) with a fresh seed instead of simply replace with a new seed.

It is more that it combines the step to fill a new output buffer with the bookkeeping when to reseed the rng. Checking and managing counters can take almost as much time as filling the output buffer with fresh values, so every little thing we have to do less each step helps.

> Not quite sure what I think right now; this would make `ReseedingRng` only usable for certain classes of RNGs, right?

I think I named it `ReseedingBlockRng`? I think nothing prevent using the current `ReseedingRng` with all kinds of RNG's. But this should be a faster alternative for what are basically `CryptoRng`. But if reseeding other kinds of RNG's does not make much sense, we could remove `ReseedingRng`.

What do you think about introducing an `RngCore` trait? In some way I find it a bit ugly because the main motivation is just to increase the performance of a reseeding wrapper. But because of `TreadRng` this may be worthwhile.

On the other hand `IsaacRng`, `ChaChaRng` and `Hc128Rng` now all have an `impl` block with the core algorithm, and this trait would make that a bit more formal. Maybe it can even become possible to share all the extra code for `impl Rng` between these implementations.