@dhardy 2017-12-18
Maybe the trait should be named `BlockRng` instead? But I think this means CSPRNGs would have to implement:

*   `Rng`
*   `CryptoRng`
*   `BlockRng`
*   `SeedableRng`
*   `SeedFromRng`

Is that a few too many traits? I'm wondering whether `Rng` should be implemented automatically for every `BlockRng` (perhaps via trait extension with default impls). Actually, if `init` is provided, then `SeedableRng` and `SeedFromRng` could be implemented automatically too? I'm not totally sure this will work, but it may simplify CSPRNG implementations.