@pitdicker 2017-12-18
Maybe there are tricks with traits I don't know, but I couldn't see a way to do this automatically. `Rng` implementations would have a struct that includes the struct from `RngCore`, an output buffer and counter/index. But a macro could certainly do it.