@pitdicker 2017-12-20
It turns out `AsRef<[T]>` is not implemented for `[u32; 256]`, so that doesn't work as a bound. @dhardy do you happen to have an idea on how to have a generic `Results` type, but still be able to do something useful like index into it or getting the length?