@dhardy 2017-12-21
Sorry, no, but I guess this is another thing that will be fixed by [constant generics](https://github.com/rust-lang/rfcs/blob/master/text/2000-const-generics.md) eventually.