@pitdicker 2017-12-24
@dhardy I now have this mostly working, and am cleaning up the changes. A _lot_ of code has to move around... Implementing RNG's with the `BlockRng` trait looks pretty clean now, here for example [HC-128](https://github.com/dhardy/rand/commit/51a555f0bc9db5a067b57e84eff9962d7fe1e7a0). The overhead of `ReseedingBlockRng` with this method is not really measurable, as hoped.

As trait I now have:
```rust
pub trait BlockRng<T>: Sized {
    type Results: AsRef<[T]> + Default;

    fn generate(&mut self, results: &mut Self::Results);
}
```

But there is one place I am stuck with the traits: https://github.com/pitdicker/rand/blob/blockrng_part2/rand_core/src/lib.rs#L168. I want to add a `BlockRngWrapper` to implement `Rng` for a `BlockRng`. And it needs a separate implementation for those that return `[u32]` and `[u64]`.

The implementation for `[u64]` is commented out at the moment, otherwise I get the error:
```
error[E0119]: conflicting implementations of trait `Rng` for type `BlockRngWrapper<_, _>
```

Do you know how to fix this?